Updated return to dashboard button and tests, moved ButtonSelect out of LegacyForms
