+++ packages/grafana-ui/src/components/index.ts
@@ -165,8 +165,10 @@ export { TextArea } from './TextArea/TextArea';
 export { FileUpload } from './FileUpload/FileUpload';
 export { TimeRangeInput } from './TimePicker/TimeRangeInput';
 export { Card, Props as CardProps, ContainerProps, CardInnerProps, getCardStyles } from './Card/Card';
-
 export { FormattedValueDisplay } from './FormattedValueDisplay/FormattedValueDisplay';
+
+export { ButtonSelect } from './Dropdown/ButtonSelect';
+
 // Legacy forms
 
 // Export this until we've figured out a good approach to inline form styles.
@@ -176,7 +178,6 @@ export { InlineFormLabel } from './FormLabel/FormLabel';
 import { Select, AsyncSelect } from './Forms/Legacy/Select/Select';
 import { IndicatorsContainer } from './Forms/Legacy/Select/IndicatorsContainer';
 import { NoOptionsMessage } from './Forms/Legacy/Select/NoOptionsMessage';
-import { ButtonSelect } from './Forms/Legacy/Select/ButtonSelect';
 
 //Input
 import { Input, LegacyInputStatus } from './Forms/Legacy/Input/Input';
@@ -192,7 +193,6 @@ const LegacyForms = {
   AsyncSelect,
   IndicatorsContainer,
   NoOptionsMessage,
-  ButtonSelect,
   Input,
   Switch,
 };
